requirements.txt
streamlit==1.41.1
google-generativeai==0.8.3
python-dotenv==1.0.1
pdf2image==1.17.0
